3-2 ArithmeticCalculator 클래스에 Generics를 통해 모든 숫자 타입 입력 가능 추가

+++ src/main/java/calculator/AddOperator.java
@@ -1,8 +1,8 @@
 package calculator;
 
-public class AddOperator implements Calculatable{
-   @Override
-    public double operate(int num1, int num2) {
+public class AddOperator implements Calculatable {
+    @Override
+    public double operate(double num1, double num2) {
         return num1 + num2; // 덧셈
     }
 }
