+++ src/main/java/calculator/DivideOperator.java
@@ -2,7 +2,7 @@ package calculator;
 
 public class DivideOperator implements Calculatable {
     @Override
-    public double operate(int num1, int num2) {
+    public double operate(double num1, double num2) {
         return (double) (num1 / num2); // 나눗셈
     }
 }
