+++ src/main/java/calculator/SubtractOperator.java
@@ -2,7 +2,7 @@ package calculator;
 
 public class SubtractOperator implements Calculatable {
     @Override
-    public double operate(int num1, int num2) {
+    public double operate(double num1, double num2) {
         return num1 - num2; // 뺄셈
     }
 }
